refactor: Separate the command line tools from the generators
