+++ ors-benchmark/src/main/java/org/heigit/ors/generators/CoordinateGeneratorSnappingApp.java
@@ -1,5 +1,6 @@
 package org.heigit.ors.generators;
 
+import org.heigit.ors.cli.CoordinateGeneratorSnappingCLI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
