+++ ors-benchmark/src/test/java/org/heigit/ors/cli/CoordinateGeneratorSnappingCLITest.java
@@ -1,14 +1,13 @@
-package org.heigit.ors.generators;
+package org.heigit.ors.cli;
 
 import org.heigit.ors.exceptions.CommandLineParsingException;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import org.heigit.ors.generators.CoordinateGeneratorSnapping;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 class CoordinateGeneratorSnappingCLITest {
